+++ src/steps/source-nodes/create-nodes/create-nodes.js
@@ -1,9 +1,6 @@
-import stringify from "fast-json-stable-stringify"
-import execall from "execall"
 import PQueue from "p-queue"
 import fetchReferencedMediaItemsAndCreateNodes from "../fetch-nodes/fetch-referenced-media-items"
 import urlToPath from "~/utils/url-to-path"
-import { getGatsbyApi } from "~/utils/get-gatsby-api"
 import store from "~/store"
 import fetchGraphql from "~/utils/fetch-graphql"
 
@@ -11,6 +8,7 @@ import {
   buildTypeName,
   getTypeSettingsByType,
 } from "~/steps/create-schema-customization/helpers"
+import { processNode } from "./process-node"
 
 // @todo concurrency is currently set so low because side effects can overwhelm
 // the remote server. A queue for the entire source plugin should be created so that
@@ -18,8 +16,6 @@ import {
 const createNodesQueue = new PQueue({
   concurrency: 2,
 })
-
-// const imgSrcRemoteFileRegex = /(?:src=\\")((?:(?:https?|ftp|file):\/\/|www\.|ftp\.)(?:\([-A-Z0-9+&@#/%=~_|$?!:,.]*\)|[-A-Z0-9+&@#/%=~_|$?!:,.])*(?:\([-A-Z0-9+&@#/%=~_|$?!:,.]*\)|[A-Z0-9+&@#/%=~_|$])\.(?:jpeg|jpg|png|gif|ico|pdf|doc|docx|ppt|pptx|pps|ppsx|odt|xls|psd|mp3|m4a|ogg|wav|mp4|m4v|mov|wmv|avi|mpg|ogv|3gp|3g2|svg|bmp|tif|tiff|asf|asx|wm|wmx|divx|flv|qt|mpe|webm|mkv|txt|asc|c|cc|h|csv|tsv|ics|rtx|css|htm|html|m4b|ra|ram|mid|midi|wax|mka|rtf|js|swf|class|tar|zip|gz|gzip|rar|7z|exe|pot|wri|xla|xlt|xlw|mdb|mpp|docm|dotx|dotm|xlsm|xlsb|xltx|xltm|xlam|pptm|ppsm|potx|potm|ppam|sldx|sldm|onetoc|onetoc2|onetmp|onepkg|odp|ods|odg|odc|odb|odf|wp|wpd|key|numbers|pages))(?=\\"| |\.)/gim
 
 const createNodeWithSideEffects = ({
   node,
@@ -32,36 +28,22 @@ const createNodeWithSideEffects = ({
   createdNodeIds,
   createNodesActivity,
   totalSideEffectNodes,
+  wpUrl,
 }) => async () => {
   if (node.link) {
+    // @todo is this still necessary? I don't think it is but double check
     // create a pathname for the node using the WP permalink
     node.path = urlToPath(node.link)
   }
 
-  // here we're searching for file strings in our node
-  // we use this to download only the media items
-  // that are being used in posts
-  // this is important not only for downloading only used images
-  // but also for downloading images in post content
   if (wpgqlNodesGroup.plural !== `mediaItems`) {
-    const nodeString = stringify(node)
-
-    // const imageUrlMatches = execall(imgSrcRemoteFileRegex, nodeString)
-
-    // if (imageUrlMatches.length) {
-    //   store.dispatch.imageNodes.addImgMatches(imageUrlMatches)
-    // }
-
-    if (!pluginOptions.type.MediaItem.lazyNodes) {
-      // get an array of all referenced media file ID's
-      const matchedIds = execall(/"id":"([^"]*)","sourceUrl"/gm, nodeString)
-        .map((match) => match.subMatches[0])
-        .filter((id) => id !== node.id)
-
-      if (matchedIds.length) {
-        matchedIds.forEach((id) => referencedMediaItemNodeIds.add(id))
-      }
-    }
+    node = await processNode({
+      node,
+      pluginOptions,
+      referencedMediaItemNodeIds,
+      wpUrl,
+      helpers,
+    })
   }
 
   const remoteNode = {
@@ -117,27 +99,9 @@ export const createGatsbyNodesFromWPGQLContentNodes = async ({
   wpgqlNodesByContentType,
   createNodesActivity,
 }) => {
-  const { helpers, pluginOptions } = getGatsbyApi()
-
-  const {
-    data: {
-      generalSettings: { url: wpUrl },
-    },
-  } = await fetchGraphql({
-    query: /* GraphQL */ `
-      query {
-        generalSettings {
-          url
-        }
-      }
-    `,
-  })
-
-  const anchorTagRegex = new RegExp(
-    // eslint-disable-next-line no-useless-escape
-    `<a[\\\s]+[^>]*?href[\\\s]?=["'\\\\]*(${wpUrl}.*?)["'\\\\]*.*?>([^<]+|.*?)?<\/a>`,
-    `gim`
-  )
+  const state = store.getState()
+  const { wpUrl } = state.remoteSchema
+  const { helpers, pluginOptions } = state.gatsbyApi
 
   // wp supports these file extensions
   // jpeg|jpg|png|gif|ico|pdf|doc|docx|ppt|pptx|pps|ppsx|odt|xls|psd|mp3|m4a|ogg|wav|mp4|m4v|mov|wmv|avi|mpg|ogv|3gp|3g2|svg|bmp|tif|tiff|asf|asx|wm|wmx|divx|flv|qt|mpe|webm|mkv|txt|asc|c|cc|h|csv|tsv|ics|rtx|css|htm|html|m4b|ra|ram|mid|midi|wax|mka|rtf|js|swf|class|tar|zip|gz|gzip|rar|7z|exe|pot|wri|xla|xlt|xlw|mdb|mpp|docm|dotx|dotm|xlsm|xlsb|xltx|xltm|xlam|pptm|ppsm|potx|potm|ppam|sldx|sldm|onetoc|onetoc2|onetmp|onepkg|odp|ods|odg|odc|odb|odf|wp|wpd|key|numbers|pages
@@ -145,7 +109,13 @@ export const createGatsbyNodesFromWPGQLContentNodes = async ({
   // gatsby-image supports these file types
   // const imgSrcRemoteFileRegex = /<img.*?src=\\"(.*?jpeg|jpg|png|webp|tif|tiff$)\\"[^>]+>/gim
 
-  const { actions, createContentDigest } = helpers
+  const { actions, createContentDigest, reporter } = helpers
+
+  store.dispatch.logger.createActivityTimer({
+    typeName: `MediaItem`,
+    pluginOptions,
+    reporter,
+  })
 
   const createdNodeIds = []
   const totalSideEffectNodes = []
@@ -167,6 +137,7 @@ export const createGatsbyNodesFromWPGQLContentNodes = async ({
           createdNodeIds,
           createNodesActivity,
           totalSideEffectNodes,
+          wpUrl,
         })
       )
     }
@@ -188,8 +159,16 @@ export const createGatsbyNodesFromWPGQLContentNodes = async ({
       referencedMediaItemNodeIds: referencedMediaItemNodeIdsArray,
     })
 
+    store.dispatch.logger.stopActivityTimer({
+      typeName: `MediaItem`,
+    })
+
     return [...createdNodeIds, ...referencedMediaItemNodeIdsArray]
   }
+
+  store.dispatch.logger.stopActivityTimer({
+    typeName: `MediaItem`,
+  })
 
   return createdNodeIds
 }